working on user login func and need a lot of refactoring

+++ frontend/src/services/apiCategory.ts
@@ -1,22 +1,37 @@
 import { CATEGORY_ID_URL, CATEGORY_URL } from "../statics/fetchUrls";
 import { NewCategory } from "../types/flashcardTypes";
+import { findToken } from "../utils/apiHelpers";
 
 export async function getCategories() {
+  const accessToken = findToken();
+  console.log("token", accessToken);
+  if (!accessToken) return alert("Please check in first");
   try {
-    const res = await fetch(CATEGORY_URL);
+    const res = await fetch(CATEGORY_URL, {
+      method: "GET",
+      headers: {
+        Authorization: accessToken,
+        "Content-Type": "application/json",
+      },
+    });
     const data = await res.json();
+    console.log(data);
     if (!data) return console.error("something went wrong with category fetching 💥");
     return data;
   } catch (err) {
+    console.log(err);
     throw new Error("Couldn't get categories");
   }
 }
 
 export async function createCategory(body: NewCategory) {
+  const accessToken = findToken();
+
+  if (!accessToken) return alert("Please check in first");
   try {
     const res = await fetch(CATEGORY_URL, {
       method: "POST",
-      headers: { "Content-Type": "application/json" },
+      headers: { Authorization: accessToken, "Content-Type": "application/json" },
       body: JSON.stringify(body),
     });
 
@@ -29,9 +44,12 @@ export async function createCategory(body: NewCategory) {
 }
 
 export async function deleteCategory(id: number | string) {
+  const accessToken = findToken();
+  if (!accessToken) return alert("Please check in first");
   try {
     const res = await fetch(CATEGORY_ID_URL(id), {
       method: "DELETE",
+      headers: { Authorization: accessToken },
     });
     const data = await res.json();
     if (!data) return console.error("something went wrong with a category deleting💥");
@@ -42,10 +60,12 @@ export async function deleteCategory(id: number | string) {
 }
 
 export async function updateCategory(id: number | string, body: { category: string }) {
+  const accessToken = findToken();
+  if (!accessToken) return alert("Please check in first");
   try {
     const res = await fetch(CATEGORY_ID_URL(id), {
       method: "PUT",
-      headers: { "Content-Type": "application/json" },
+      headers: { Authorization: accessToken, "Content-Type": "application/json" },
       body: JSON.stringify(body),
     });
