+++ frontend/src/services/apiDeck.ts
@@ -6,11 +6,20 @@ import {
   DECK_WITH_CATEGOY_URL,
   DECK_WITH_DATE_CATEGOY_URL,
 } from "../statics/fetchUrls";
-import { DeckType, NewDeckType } from "../types/flashcardTypes";
+import { NewDeckType } from "../types/flashcardTypes";
+import { findToken } from "../utils/apiHelpers";
 
 export async function getAllDecks() {
+  const accessToken = findToken();
+  if (!accessToken) return alert("Please check in first");
   try {
-    const res = await fetch(fetchAllDecksUrl);
+    const res = await fetch(fetchAllDecksUrl, {
+      method: "GET",
+      headers: {
+        Authorization: accessToken,
+        "Content-Type": "application/json",
+      },
+    });
     const data = await res.json();
     if (!data) return console.error("something went wrong with all decks fetching 💥");
     return data;
@@ -20,8 +29,17 @@ export async function getAllDecks() {
 }
 
 export async function getDecksWithQuery(query: string) {
+  const accessToken = findToken();
+  if (!accessToken) return alert("Please check in first");
+
   try {
-    const res = await fetch(fetchDecksWithQuery(query));
+    const res = await fetch(fetchDecksWithQuery(query), {
+      method: "GET",
+      headers: {
+        Authorization: accessToken,
+        "Content-Type": "application/json",
+      },
+    });
     const data = await res.json();
     if (!data) return console.error("something went wrong with decks fetching 💥");
     return data;
@@ -31,8 +49,16 @@ export async function getDecksWithQuery(query: string) {
 }
 
 export async function getDecksWithCategopry(categoryId: string, query: string) {
+  const accessToken = findToken();
+  if (!accessToken) return alert("Please check in first");
   try {
-    const res = await fetch(DECK_WITH_CATEGOY_URL(categoryId, query));
+    const res = await fetch(DECK_WITH_CATEGOY_URL(categoryId, query), {
+      method: "GET",
+      headers: {
+        Authorization: accessToken,
+        "Content-Type": "application/json",
+      },
+    });
     const data = await res.json();
     if (!data) return console.error("something went wrong with decks fetching 💥");
     return data;
@@ -42,8 +68,17 @@ export async function getDecksWithCategopry(categoryId: string, query: string) {
 }
 
 export async function getDatesOfDecks(categoryId: string, query: string) {
+  if (!categoryId) return;
+  const accessToken = findToken();
+  if (!accessToken) return alert("Please check in first");
   try {
-    const res = await fetch(DECK_WITH_DATE_CATEGOY_URL(categoryId, query));
+    const res = await fetch(DECK_WITH_DATE_CATEGOY_URL(categoryId, query), {
+      method: "GET",
+      headers: {
+        Authorization: accessToken,
+        "Content-Type": "application/json",
+      },
+    });
     const data = await res.json();
     if (!data) return console.error("something went wrong with deck's dates fetching 💥");
     return data;
@@ -54,8 +89,16 @@ export async function getDatesOfDecks(categoryId: string, query: string) {
 
 export async function getDeck(id: number | string | undefined) {
   if (id === undefined) return;
+  const accessToken = findToken();
+  if (!accessToken) return alert("Please check in first");
   try {
-    const res = await fetch(fetchDeckByIdUrl(id));
+    const res = await fetch(fetchDeckByIdUrl(id), {
+      method: "GET",
+      headers: {
+        Authorization: accessToken,
+        "Content-Type": "application/json",
+      },
+    });
     const data = await res.json();
     if (!data) return console.error("something went wrong with a deck fetching 💥");
     return data;
@@ -65,10 +108,12 @@ export async function getDeck(id: number | string | undefined) {
 }
 
 export async function createDeck(body: NewDeckType) {
+  const accessToken = findToken();
+  if (!accessToken) return alert("Please check in first");
   try {
     const res = await fetch(createDeckUrl, {
       method: "POST",
-      headers: { "Content-Type": "application/json" },
+      headers: { Authorization: accessToken, "Content-Type": "application/json" },
       body: JSON.stringify(body),
     });
 
@@ -81,9 +126,12 @@ export async function createDeck(body: NewDeckType) {
 }
 
 export async function deleteDeck(id: number | string) {
+  const accessToken = findToken();
+  if (!accessToken) return alert("Please check in first");
   try {
     const res = await fetch(fetchDeckByIdUrl(id), {
       method: "DELETE",
+      headers: { Authorization: accessToken },
     });
     const data = await res.json();
     if (!data) return console.error("something went wrong with a deck deleting💥");
@@ -94,10 +142,12 @@ export async function deleteDeck(id: number | string) {
 }
 
 export async function updateDeck(id: number | string, body: NewDeckType) {
+  const accessToken = findToken();
+  if (!accessToken) return alert("Please check in first");
   try {
     const res = await fetch(fetchDeckByIdUrl(id), {
       method: "PUT",
-      headers: { "Content-Type": "application/json" },
+      headers: { Authorization: accessToken, "Content-Type": "application/json" },
       body: JSON.stringify(body),
     });
