+++ frontend/src/services/apiFlashcard.ts
@@ -1,9 +1,18 @@
 import { createFlashcardUrl, fetchFlashcardByIdUrl } from "../statics/fetchUrls";
 import { CardType, CreateFlashCard } from "../types/flashcardTypes";
+import { findToken } from "../utils/apiHelpers";
 
 export async function getFlashCard(id: number | string) {
+  const accessToken = findToken();
+  if (!accessToken) return alert("Please check in first");
   try {
-    const res = await fetch(fetchFlashcardByIdUrl(id));
+    const res = await fetch(fetchFlashcardByIdUrl(id), {
+      method: "GET",
+      headers: {
+        Authorization: accessToken,
+        "Content-Type": "application/json",
+      },
+    });
     const data = await res.json();
     if (!data) return console.error("something went wrong with a flashcard fetching 💥");
     return data;
@@ -13,10 +22,12 @@ export async function getFlashCard(id: number | string) {
 }
 
 export async function createFlashCard(body: CreateFlashCard) {
+  const accessToken = findToken();
+  if (!accessToken) return alert("Please check in first");
   try {
     const res = await fetch(createFlashcardUrl, {
       method: "POST",
-      headers: { "Content-Type": "application/json" },
+      headers: { Authorization: accessToken, "Content-Type": "application/json" },
       body: JSON.stringify(body),
     });
     const data = await res.json();
@@ -30,9 +41,12 @@ export async function createFlashCard(body: CreateFlashCard) {
 export async function deleteFlashCard(id: number | string | undefined) {
   console.log(id);
   if (id === undefined) return;
+  const accessToken = findToken();
+  if (!accessToken) return alert("Please check in first");
   try {
     const res = await fetch(fetchFlashcardByIdUrl(id), {
       method: "DELETE",
+      headers: { Authorization: accessToken },
     });
     const data = await res.json();
     console.log(data);
@@ -45,10 +59,12 @@ export async function deleteFlashCard(id: number | string | undefined) {
 
 export async function updateFlashCard(id: string | undefined, body: CardType | { status: string }) {
   if (id === undefined) return;
+  const accessToken = findToken();
+  if (!accessToken) return alert("Please check in first");
   try {
     const res = await fetch(fetchFlashcardByIdUrl(id), {
       method: "PUT",
-      headers: { "Content-Type": "application/json" },
+      headers: { Authorization: accessToken, "Content-Type": "application/json" },
       body: JSON.stringify(body),
     });
